reorganized folders, removed unused imports
