+++ src/components/Button/Button.tsx
@@ -1,4 +1,4 @@
-import React, { MouseEvent } from "react";
+import React from "react";
 import styles from "../Button/Button.module.css";
