+++ src/pages/_app.tsx
@@ -1,6 +1,4 @@
 import '../styles/globals.css'
-import NavBar from '../components/NavBarAnd/NavBar';
-import IntakeFormOverlay from '../components/IntakeFormOverlay/IntakeFormOverlay';
 
 function MyApp({ Component, pageProps }) {
   return (
