+++ src/pages/calendar/index.tsx
@@ -1,5 +1,4 @@
-import Link from 'next/link'
-import Layout from '../../components/Layout'
+import Layout from '../../components/layout'
 
 export default function Calendar() {
     return (
