+++ src/pages/index.tsx
@@ -1,11 +1,6 @@
-import Head from 'next/head'
-import styles from '../styles/Home.module.css'
 import firebase from '../firebase/clientApp'
 import { useCollection } from "react-firebase-hooks/firestore";
-import { TextRegular, TextRegularBold } from '../public/assets/fonts/Fonts';
-import Question from '../components/Question/question';
 import Dashboard from './dashboard';
-import GenQuestion from './IntakeForms/GenQuestion';
 import Layout from '../components/layout';
 
 export default function Home() {
